requirements.txt
# Core
pytz==2020.1
Django==3.0.8
django-configurations==2.2
gunicorn==20.0.4
newrelic==5.14.1.144

# For the persistence stores
psycopg2-binary==2.8.5
dj-database-url==0.5.0
mysqlclient==2.0.1

# Model Tools
django-model-utils==4.0.0
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.11.0
Markdown==3.2.2
django-filter==2.3.0

# Developer Tools
ipdb==0.13.3
ipython==7.16.1
mkdocs==1.1.2
flake8==3.8.3

# Testing
mock==4.0.2
factory-boy==2.12.0
django-nose==1.4.6
nose-progressive==1.5.2
coverage==5.2

# Static and Media Storage
django-storages==1.9.1
boto3==1.14.20

